chore(suite): moved unpair list to the parent
